Split settings panel sections
