+++ apps/desktop/src/renderer/components/sidebar/settings-panel-styles.ts
@@ -1,0 +1,4 @@
+export const inputCls = 'w-full px-3 py-2 rounded-lg text-[12px] bg-base border border-border-subtle text-text placeholder:text-text-muted outline-none focus:border-accent/40 transition-colors'
+export const sectionLabelCls = 'text-[10px] font-semibold uppercase tracking-widest text-text-muted px-1'
+export const fieldLabelCls = 'text-[11px] text-text-muted font-medium'
+export const panelCardCls = 'rounded-2xl border border-border-subtle p-4 flex flex-col gap-4'
